refactor user model and moved google env variables to google oauth route from google utils
