+++ server/api/auth/auth.ts
@@ -1,6 +1,6 @@
-import { User, UserAttributes } from '../../database/models/user'
 import express, { NextFunction, Request, Response } from 'express'
 
+import { User } from '../../database/index'
 import { getGitHubInfo } from './Utils/GithubUtils'
 import { getGoogleInfo } from './Utils/GoogleUtils';
 
@@ -10,18 +10,18 @@ router.get("/", async (req: Request, res: Response, next: NextFunction) => {
   try{
     const { googleAccessToken, gitHubAccessToken } = req.cookies
 
-    let user: Partial<UserAttributes> = {}
+    let user: Partial<typeof User>|null = {}
     
     if(googleAccessToken) {
       const { email }: { email: string, name: string } = (await getGoogleInfo(googleAccessToken)).data
-      user = await User.findUser({
+      user = await User.findOne({
         where: {
           email: email,
         },
       });
     }else if (gitHubAccessToken) {
       const { login: username } = (await getGitHubInfo(gitHubAccessToken)).data;
-      user = await User.findUser({
+      user = await User.findOne({
         where: {
           username,
         },
