+++ server/api/users.ts
@@ -1,14 +1,12 @@
 import express, { NextFunction, Request, Response } from 'express'
 
-import { User } from '../database/models/user';
-import db from '../database/models';
+import { User } from '../database/index';
 
 const router = express.Router();
 
 router.get('/', async (req: Request, res: Response, next: NextFunction) => {
   try{
-    console.log(req.cookies)
-    const users = await User.findAllUser()
+    const users = await User.findAll()
     res.send(users)
   } catch(err) {
     next(err)
