+++ server/database/index.ts
@@ -1,0 +1,9 @@
+import db from './models/index'
+
+export const {
+  User,
+  Problem,
+  ProblemInfo
+} = db 
+
+export default db
